Reorganise pages and router folders
