+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
-import { router } from "./components/router";
+import { router } from "./router";
 import "@fontsource/ubuntu/700.css";
 import "@fontsource/ubuntu/400.css";
 import "@fontsource/ubuntu/500.css";
