+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,4 +1,4 @@
-import Button from "../../Button/Button";
+import Button from "../../components/Button/Button";
 import "./NotFoundPage.css";
 
 const NotFoundPage: React.FC = () => {
@@ -10,11 +10,7 @@ const NotFoundPage: React.FC = () => {
           This page doesn't exist. Not to worry, it happens to even the best
           riders & teams.
         </p>
-        <Button
-          text={"Go Back Home"}
-          redirection={"/home"}
-          className="button"
-        />
+        <Button text="Go Back Home" redirection="/home" className="button" />
       </div>
     </>
   );
